Rename `PageEntityTable` to `PaginatedEntityTable`.
