+++ graylog2-web-interface/src/components/common/index.tsx
@@ -83,7 +83,7 @@ export { default as ModalSubmit } from './ModalSubmit';
 export { default as MultiSelect } from './MultiSelect';
 export { default as NestedForm } from './NestedForm';
 export { default as OverlayTrigger } from './OverlayTrigger';
-export { default as PageEntityTable } from './PageEntityTable/PageEntityTable';
+export { default as PaginatedEntityTable } from 'components/common/PaginatedEntityTable/PaginatedEntityTable';
 export { default as PageErrorOverview } from './PageErrorOverview';
 export { default as PageHeader } from './PageHeader';
 export { default as PageSizeSelect } from './PageSizeSelect';
