+++ graylog2-web-interface/src/components/indices/IndexSetFieldTypes/IndexSetFieldTypesList.test.tsx
@@ -22,7 +22,7 @@ import { ReactRouter6Adapter } from 'use-query-params/adapters/react-router-6';
 import { MockStore } from 'helpers/mocking';
 import useParams from 'routing/useParams';
 import asMock from 'helpers/mocking/AsMock';
-import useFetchEntities from 'components/common/PageEntityTable/useFetchEntities';
+import useFetchEntities from 'components/common/PaginatedEntityTable/useFetchEntities';
 import useUserLayoutPreferences from 'components/common/EntityDataTable/hooks/useUserLayoutPreferences';
 import { layoutPreferences } from 'fixtures/entityListLayoutPreferences';
 import TestStoreProvider from 'views/test/TestStoreProvider';
@@ -73,7 +73,7 @@ jest.mock('stores/indices/IndexSetsStore', () => ({
 
 jest.mock('routing/useParams', () => jest.fn());
 jest.mock('views/logic/fieldactions/ChangeFieldType/hooks/useFieldTypesForMappings');
-jest.mock('components/common/PageEntityTable/useFetchEntities');
+jest.mock('components/common/PaginatedEntityTable/useFetchEntities');
 
 jest.mock('components/common/EntityDataTable/hooks/useUserLayoutPreferences');
 jest.mock('components/indices/IndexSetFieldTypeProfiles/hooks/useProfile');
